delete: Check pod state before to issue a StopPod
The way a delete works, it was always trying to stop the pod, even
when the force flag was not enabled. Because we want to be able to
stop the pod from a kill command, this means a pod stop might be
called twice, and we don't want the second stop to fail, leading
to the failure of the delete command.

That's why this commit checks for the pod status before to try
stopping it.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>